Adding TODO.
Now that I got a physical prototype everything worked out perfect except
for the 5mm hole sizes and it's more than expected flimsyness.

The hole problem is because the powdercoat slightly shrunk the hole
diameter, but I worked around this by using a mallet to drive a peg
through the hole. Ended up fitting snugly as I initially expected
afterwards, so that problem is solved.

However 2 mm of thickness is more flimsy than I thought. If The UPS's
rails were shoter I would mount the UPS on the rack, and I would feel
better if it were a bit thicker. I read somewhere that racks should be
2mm to 3mm thick, so going with the low end was a small mistake. However
since I wont be mounting the UPS on the rack there won't be as much
weight bearing down on the material, so 2 mm works fine for my current
use case.

No extended use cases are planned. Probably within a year I'll move back
to Austin TX and buy property again with enough room for real racks.